// src/app/business-logic/model/users/usersSetPreferencesRequest.ts
/**
 * users
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.7
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface UsersSetPreferencesRequest {
    /**
     * Updates to user preferences. A mapping from keys in dot notation to values. For   example, `{\"a.b\": 0}` will set the key \"b\" in object \"a\" to 0.
     */
    preferences: object;
}